requirements.txt
asgiref==3.7.2
beautifulsoup4==4.10.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.2.3
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.10.0
django-bootstrap4==23.2
django-crispy-forms==1.14.0
django-filter==23.2
django-import-export==2.7.1
django-tables2==2.6.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
PyYAML==6.0
soupsieve==2.3.1
sqlparse==0.4.2
tablib==3.2.0
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
